+++ kaliphonestudio/__init__.py
@@ -1,1 +1,1 @@
-__version__ = "0.6.64-dev"
+__version__ = "0.6.65-dev"